move document section around
