--- src/document/parse/postProcess/index.js
@@ -1,12 +1,0 @@
-//cleanup after ourselves..
-const postProcess = function(r) {
-  //try to guess the page's title (from the bold first-line)
-  if (r.sections[0] && r.sections[0].sentences[0]) {
-    let s = r.sections[0].sentences[0];
-    if (s.fmt && s.fmt.bold && s.fmt.bold[0]) {
-      r.title = r.title || s.fmt.bold[0];
-    }
-  }
-  return r;
-};
-module.exports = postProcess;
