+++ src/document/preProcess/coordinates.js
